Enhance video processing in video.py: Added logging for available ONNX providers, updated colorizer model usage, and improved audio handling by checking for audio streams before remuxing. Cleaned up commented-out code for better readability.

+++ video.py
@@ -1,4 +1,3 @@
-
 import os
 import sys
 import cv2
@@ -11,6 +10,8 @@ from tqdm import tqdm
 
 import onnxruntime as rt
 rt.set_default_logger_severity(3)
+print("Available Providers:", rt.get_available_providers())
+print("CUDA Available:", 'CUDAExecutionProvider' in rt.get_available_providers())
 
 parser = ArgumentParser()
 parser.add_argument("--source", help="path to source video")
@@ -39,10 +40,10 @@ render_factor = opt.render_factor * 32
 #colorizer = DEOLDIFY(model_path="color/deoldify.onnx", device="cuda")
 
 # new onnx models - render_factor - dynamic axes input:
-from color.deoldify_fp16 import DEOLDIFY
-colorizer = DEOLDIFY(model_path="color/ColorizeArtistic_dyn_fp16.onnx", device="cuda")
-#from color.deoldify import DEOLDIFY
-#colorizer = DEOLDIFY(model_path="color/ColorizeArtistic_dyn.onnx", device="cuda")
+# from color.deoldify import DEOLDIFY
+# colorizer = DEOLDIFY(model_path="color/ColorizeArtistic_dyn_fp16.onnx", device="cuda")
+from color.deoldify import DEOLDIFY
+colorizer = DEOLDIFY(model_path="color/ColorizeArtistic_dyn.onnx", device="cuda")
 
 
 video = cv2.VideoCapture(opt.source)
@@ -66,15 +67,27 @@ for frame_idx in tqdm(range(n_frames)):
     result = colorizer.colorize(frame, render_factor)
     
     writer.write(result)
-    cv2.imshow ("Result",result)
-    k = cv2.waitKey(1)
-    if k == 27:
-        writer.release()
-        break
+    # cv2.imshow("Result", result)  # Comment out or remove this line
+    # k = cv2.waitKey(1)           # Comment out or remove this line
+    # if k == 27:                  # Comment out or remove this line
+    #     writer.release()         # Comment out or remove this line
+    #     break                    # Comment out or remove this line
+
+# Add these lines to properly close the video objects
+writer.release()
+video.release()
 
 if opt.audio:
-    # lossless remuxing audio/video
-    command = 'ffmpeg.exe -y -vn -i ' + '"' + opt.source + '"' + ' -an -i ' + 'temp.mp4' + ' -c:v copy -acodec libmp3lame -ac 2 -ar 44100 -ab 128000 -map 0:1 -map 1:0 -shortest ' + '"' + opt.result + '"'
-    subprocess.call(command, shell=platform.system() != 'Windows')
-    os.remove('temp.mp4')
+    # Check if source video has audio stream
+    probe_command = f'ffprobe -i "{opt.source}" -show_streams -select_streams a -loglevel error'
+    has_audio = subprocess.call(probe_command, shell=platform.system() != 'Windows') == 0
+
+    if has_audio:
+        # lossless remuxing audio/video
+        command = 'ffmpeg -y -vn -i ' + '"' + opt.source + '"' + ' -an -i ' + 'temp.mp4' + ' -c:v copy -acodec libmp3lame -ac 2 -ar 44100 -ab 128000 -map 0:1 -map 1:0 -shortest ' + '"' + opt.result + '"'
+        subprocess.call(command, shell=platform.system() != 'Windows')
+        os.remove('temp.mp4')
+    else:
+        # If no audio, just rename temp.mp4 to final result
+        os.rename('temp.mp4', opt.result)
